requirements.txt
plaid-python
requests
pandas
python-dotenv
